src/requirement.txt
asn1crypto==1.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
icc-rt==2020.0.133
idna==2.8
intel-openmp==2020.0.133
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
mkl==2019.0
mkl-fft==1.0.6
mkl-random==1.0.1.1
numpy==1.18.1
olefile==0.46
pandas==1.0.1
Pillow==7.0.0
pycparser==2.19
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.3
requests==2.22.0
six==1.14.0
tbb==2019.0
tbb4py==2019.0
urllib3==1.25.8
Werkzeug==1.0.0
